Adding Custom Criteria
Comparisons to Mirador OMI NO2d showed problems (Baublitz and
Henderson, 2013) that were the result of updated quality control
algorithms used at NASA. The quality controls are likely to be updated
again or to be specific to a particular use-case. To facilitate this,
we have added a customCriteria option that interprets NASA OMI-style QC
criteria.  The criteria as documented by OMI-NO2 can be input as a
custom criteria.  This allows for ad-hoc criteria applications.